requirements.txt
duckdb
pandas
neo4j
langchain
openai
python-dotenv
streamlit

